Fixed Test because of Refactoring
